Matching progress in WaterGun. (#101)
* Matching progress in WaterGun.

Mainly matching work in TNozzleBase::emit, TNozzleDeform::emit and TNozzleTrigger::emit.

Improved matches in various other functions in the TU.

* Changes based on PR reviews.

- Added TWaterGun::offFlag()
- Removed manual stack space padding.
- Reconciled TWaterGun::calcAnimation into one big switch statement.

+++ include/Player/ModelWaterManager.hpp
@@ -73,7 +73,7 @@ public:
 	void getWaterAlpha() const;
 	bool askHitWaterParticleOnGround(const JGeometry::TVec3<f32>&);
 	void makeEmit(const TWaterEmitInfo&);
-	int emitRequest(const TWaterEmitInfo&);
+	u8 emitRequest(const TWaterEmitInfo&);
 	void splashSound(const JGeometry::TVec3<f32>&, f32) const;
 	void splashGround(int);
 	void touchingExec(int);
